refactor(easyadmin): move admin helpers to bridge and deprecate legacy config
