+++ src/Service/AuditLogExportResponseFactory.php
@@ -4,49 +4,27 @@ declare(strict_types=1);
 
 namespace Rcsofttech\AuditTrailBundle\Service;
 
-use DateTimeImmutable;
-use RuntimeException;
-use Symfony\Component\HttpFoundation\StreamedResponse;
+use Rcsofttech\AuditTrailBundle\Bridge\EasyAdmin\Service\AuditLogExportResponseFactory as BridgeAuditLogExportResponseFactory;
 
-use function fclose;
-use function fopen;
-use function sprintf;
+use function trigger_deprecation;
 
-final readonly class AuditLogExportResponseFactory
+trigger_deprecation(
+    'rcsofttech/audit-trail-bundle',
+    '4.1',
+    'The "%s" class is deprecated since rcsofttech/audit-trail-bundle 4.1; use "%s" instead.',
+    AuditLogExportResponseFactory::class,
+    BridgeAuditLogExportResponseFactory::class,
+);
+
+/**
+ * @deprecated since rcsofttech/audit-trail-bundle 4.1, use
+ *             Rcsofttech\AuditTrailBundle\Bridge\EasyAdmin\Service\AuditLogExportResponseFactory instead.
+ */
+readonly class AuditLogExportResponseFactory extends BridgeAuditLogExportResponseFactory
 {
-    private const array CONTENT_TYPES = [
-        'json' => 'application/json',
-        'csv' => 'text/csv',
-    ];
-
     public function __construct(
-        private AuditLogAdminOperations $operations,
+        AuditLogAdminOperations $operations,
     ) {
-    }
-
-    /**
-     * @param array<string, mixed> $filters
-     */
-    public function createResponse(array $filters, string $format): StreamedResponse
-    {
-        $fileName = sprintf('audit_logs_%s.%s', new DateTimeImmutable()->format('Y-m-d_His'), $format);
-
-        $response = new StreamedResponse(function () use ($filters, $format): void {
-            $output = fopen('php://output', 'w');
-            if ($output === false) {
-                throw new RuntimeException('Failed to open output stream for export');
-            }
-
-            try {
-                $this->operations->exportToStream($filters, $format, $output);
-            } finally {
-                fclose($output);
-            }
-        });
-
-        $response->headers->set('Content-Type', self::CONTENT_TYPES[$format] ?? 'application/octet-stream');
-        $response->headers->set('Content-Disposition', sprintf('attachment; filename="%s"', $fileName));
-
-        return $response;
+        parent::__construct($operations);
     }
 }
