+++ tests/Unit/Service/AuditLogAdminViewFactoryTest.php
@@ -12,16 +12,16 @@ use EasyCorp\Bundle\EasyAdminBundle\Contracts\Router\AdminRouteGeneratorInterfac
 use EasyCorp\Bundle\EasyAdminBundle\Registry\AdminControllerRegistry;
 use EasyCorp\Bundle\EasyAdminBundle\Router\AdminUrlGenerator;
 use PHPUnit\Framework\TestCase;
+use Rcsofttech\AuditTrailBundle\Bridge\EasyAdmin\Service\AuditLogAdminLocator;
+use Rcsofttech\AuditTrailBundle\Bridge\EasyAdmin\Service\AuditLogAdminOperations;
+use Rcsofttech\AuditTrailBundle\Bridge\EasyAdmin\Service\AuditLogAdminRequestMapper;
+use Rcsofttech\AuditTrailBundle\Bridge\EasyAdmin\Service\AuditLogAdminViewFactory;
+use Rcsofttech\AuditTrailBundle\Bridge\EasyAdmin\Service\AuditLogExportResponseFactory;
 use Rcsofttech\AuditTrailBundle\Contract\AuditExporterInterface;
 use Rcsofttech\AuditTrailBundle\Contract\AuditLogRepositoryInterface;
 use Rcsofttech\AuditTrailBundle\Contract\AuditReverterInterface;
 use Rcsofttech\AuditTrailBundle\Entity\AuditLog;
 use Rcsofttech\AuditTrailBundle\Enum\AuditAction;
-use Rcsofttech\AuditTrailBundle\Service\AuditLogAdminLocator;
-use Rcsofttech\AuditTrailBundle\Service\AuditLogAdminOperations;
-use Rcsofttech\AuditTrailBundle\Service\AuditLogAdminRequestMapper;
-use Rcsofttech\AuditTrailBundle\Service\AuditLogAdminViewFactory;
-use Rcsofttech\AuditTrailBundle\Service\AuditLogExportResponseFactory;
 use Rcsofttech\AuditTrailBundle\Service\RevertPreviewFormatter;
 use Rcsofttech\AuditTrailBundle\Service\TransactionDrilldownService;
 use RuntimeException;
